tests/inputdevicecapabilities-wheel.html

<!DOCTYPE html>
<title>InputDeviceCapabilities test for wheel-triggered events</title>
<meta name="viewport" content="initial-scale=1">
<meta charset="UTF-8">
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="inputdevicecapabilities-tests.js"></script>
<style>
#target {
  border: 1px solid black;
  width: 100px;
  height: 100px;
  background-color: lightgreen;
  margin-bottom: 20px;
}
</style>

<h1>InputDeviceCapabilities - test for wheel-triggered events</h1>
<p>Mousewheel over the box below.</p>
<div id="target" tabindex=1></div>
<div><b>Expected events:</b> <span id="eventList"></span></div>
<div id="log"></div>
<script>
'use strict';

function inputDeviceCapabilitiesValidator(sourceCapabilities) {
  assert_true(sourceCapabilities != null, "sourceCapabilities is set");
  assert_false(sourceCapabilities.firesTouchEvents, "sourceCapabilities.firesTouchEvents");
}

function runTests() {
  var events = ["wheel"];
  runInputDeviceCapabilitiesEventTests(events, inputDeviceCapabilitiesValidator);
}
</script>
